Extract `manually_close_market` tests
