+++ zrml/prediction-markets/src/tests/mod.rs
@@ -24,6 +24,7 @@ mod approve_market;
 mod buy_complete_set;
 mod create_market;
 mod edit_market;
+mod manually_close_market;
 mod on_market_close;
 mod reject_market;
 mod sell_complete_set;
